Linux fixes

+++ Sources/RubyBridge/RbVM.swift
@@ -45,7 +45,7 @@ final class RbVM {
 
         // Paranoid about reentrant symbol lookup during finalizers...
         var attr = pthread_mutexattr_t()
-        pthread_mutexattr_settype(&attr, PTHREAD_MUTEX_RECURSIVE)
+        pthread_mutexattr_settype(&attr, Int32(PTHREAD_MUTEX_RECURSIVE))
         pthread_mutex_init(&mutex, &attr)
     }
